confirm: Move to interactive package
Confirmation prompts are ultimately interactive, so moving it inside the
interactive package removes unnecessary clutter.
