+++ cmd/install/cmd.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/openshift/rosa/cmd/install/addon"
 	"github.com/openshift/rosa/pkg/arguments"
-	"github.com/openshift/rosa/pkg/confirm"
 	"github.com/openshift/rosa/pkg/interactive"
+	"github.com/openshift/rosa/pkg/interactive/confirm"
 )
 
 var Cmd = &cobra.Command{
